Fixes for better uint16 type

+++ src/ape_socket.h
@@ -257,7 +257,7 @@ int APE_socket_writev(ape_socket *socket, const struct iovec *iov, int iovcnt);
 int APE_sendfile(ape_socket *socket, const char *file);
 int APE_socket_is_online(ape_socket *socket);
 
-int APE_socket_port(ape_socket *socket);
+uint16_t APE_socket_port(ape_socket *socket);
 char *APE_socket_ipv4(ape_socket *socket);
 
 int ape_socket_destroy(ape_socket *socket);
